refactor(memory_bridge_adapter/store): split tests to sibling tests_store.rs (#1266)
Inline test module in store.rs is moved into a sibling tests_store.rs
following the existing memory_bridge_adapter convention (already has
tests.rs, bridge_tests.rs, test_helpers.rs).

- src/memory_bridge_adapter/store.rs: 419 → 349 LOC (under 400-line rule)
- src/memory_bridge_adapter/tests_store.rs: new (~67 LOC)
- src/memory_bridge_adapter/mod.rs: register mod tests_store
- super::* → super::store::* (sibling level)
- Added MemoryStore trait import (test methods come via trait)

Pure refactor: zero logic changes, all 7 tests pass.

Closes 2/70 of #1266 (this iteration), 9 PRs cumulative since #1285.
